@@ -1,11 +1,11 @@
 name: create_atom
 description: An animated widget that displays atom with three electrons revolving around a nucleus at its center.
-version: 2.0.0
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/Abhi011999/create_atom_flutter
 repository: https://github.com/Abhi011999/create_atom_flutter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
